requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2023.1
Django==4.2.10
django-crispy-forms==2.1
Django-Verify-Email==2.0.3
pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0